Bug 467193 - Moved files to wizard package
