+++ org.eclipse.buildship.ui/src/main/java/org/eclipse/buildship/ui/wizard/project/AbstractWizardPage.java
@@ -9,7 +9,7 @@
  *     Etienne Studer & Donát Csikós (Gradle Inc.) - initial API and implementation and initial documentation
  */
 
-package org.eclipse.buildship.ui.projectimport;
+package org.eclipse.buildship.ui.wizard.project;
 
 import java.util.List;
