+++ org.eclipse.buildship.ui/src/main/java/org/eclipse/buildship/ui/wizard/project/GradleWelcomeWizardPage.java
@@ -9,7 +9,7 @@
  *     Etienne Studer & Donát Csikós (Gradle Inc.) - initial API and implementation and initial documentation
  */
 
-package org.eclipse.buildship.ui.projectimport;
+package org.eclipse.buildship.ui.wizard.project;
 
 import com.google.common.collect.ImmutableList;
